pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opablt"
dependencies = ["attrs"]
requires-python = ">= 3.8"
version = "2024.05.01"
